make calculateBill an external function
make calculateBill an external function , in middeleware dir
