+++ server/services/order.js
@@ -1,17 +1,7 @@
 const Order = require('../models/order');
 const Product = require('../models/product'); 
+const calculateBill = require('../middleware/calculateBill'); 
 
-
-const calculateBill = async (items) => {
-    let total = 0;
-    for (const item of items) {
-        const product = await Product.findById(item.id);
-        if (product) {
-            total += product.price * item.quantity;
-        }
-    }
-    return total;
-};
 
 const createOrder = async (customerId, items, ordered, bill) => {
